requirements.txt
altgraph==0.17.5
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==12.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pystray==0.19.5
python-xlib==0.33
setuptools==80.9.0
six==1.17.0
